Swictched links and routes from names to ids to handle duplicated artists
