+++ app/api/library.py
@@ -313,11 +313,14 @@ async def get_artists(
 
 @router.get("/api/albums")
 async def get_albums(
-    artist: str = None, album_mbid: str = None, db: asyncpg.Connection = Depends(get_db)
+    artist: str = None,
+    artist_mbid: str = None,
+    album_mbid: str = None,
+    db: asyncpg.Connection = Depends(get_db),
 ):
     # 1. If artist is provided, find their MBID to classify 'main' vs 'appears_on'
-    target_mbid = None
-    if artist:
+    target_mbid = artist_mbid
+    if artist and not target_mbid:
         row = await db.fetchrow(
             """
             SELECT mbid 
@@ -336,6 +339,7 @@ async def get_albums(
             MAX(t.artwork_id) as artwork_id,
             MAX(a.sha1) as art_sha1,
             COALESCE(MAX(t.album_artist), MAX(t.artist)) as artist_name,
+            MAX(t.album_artist_mbid) as artist_mbid,
             MAX(CASE WHEN t.bit_depth > 16 OR t.sample_rate_hz > 44100 THEN 1 ELSE 0 END) as is_hires,
             MIN(t.release_date) as year,
             COUNT(DISTINCT t.id) as track_count,
@@ -365,8 +369,17 @@ async def get_albums(
     param_num = 2
 
     filters = []
-    if artist:
-        # Filter by any artist associated with the tracks via track_artists
+    if target_mbid:
+        # Filter by specific artist MBID
+        query += """
+            JOIN track_artist ta ON t.id = ta.track_id
+            JOIN artist ar ON ta.artist_mbid = ar.mbid
+        """
+        filters.append(f"ta.artist_mbid = ${param_num}")
+        params.append(target_mbid)
+        param_num += 1
+    elif artist:
+        # Filter by artist name (fallback)
         query += """
             JOIN track_artist ta ON t.id = ta.track_id
             JOIN artist ar ON ta.artist_mbid = ar.mbid
@@ -444,6 +457,9 @@ async def get_tracks(
             t.release_mbid,
             t.release_group_mbid,
             t.artwork_id,
+            t.artist_mbid,
+            t.album_artist_mbid,
+            t.release_group_mbid as album_mbid,
             a.sha1 as art_sha1,
             (SELECT STRING_AGG(a2.name, ', ' ORDER BY a2.name) 
              FROM track_artist ta2 
@@ -526,6 +542,8 @@ async def get_new_releases(limit: int = 20, db: asyncpg.Connection = Depends(get
             COUNT(DISTINCT t.id) as track_count,
             SUM(t.duration_seconds) as total_duration,
             MAX(t.release_mbid) as release_mbid,
+            MAX(t.release_group_mbid) as album_mbid,
+            (SELECT mbid FROM artist WHERE name = COALESCE(MAX(t.album_artist), MAX(t.artist)) LIMIT 1) as artist_mbid,
             'main' as type
         FROM track t
         LEFT JOIN artwork a ON t.artwork_id = a.id
@@ -553,6 +571,8 @@ async def get_recently_added_albums(
             COUNT(DISTINCT t.id) as track_count,
             SUM(t.duration_seconds) as total_duration,
             MAX(t.release_mbid) as release_mbid,
+            MAX(t.release_group_mbid) as album_mbid,
+            (SELECT mbid FROM artist WHERE name = COALESCE(MAX(t.album_artist), MAX(t.artist)) LIMIT 1) as artist_mbid,
             'main' as type
         FROM track t
         LEFT JOIN artwork a ON t.artwork_id = a.id
@@ -581,6 +601,8 @@ async def get_recently_played_albums(
             COUNT(DISTINCT t.id) as track_count,
             SUM(t.duration_seconds) as total_duration,
             MAX(t.release_mbid) as release_mbid,
+            MAX(t.release_group_mbid) as album_mbid,
+            (SELECT mbid FROM artist WHERE name = COALESCE(MAX(t.album_artist), MAX(t.artist)) LIMIT 1) as artist_mbid,
             MAX(ph.timestamp) as last_played
         FROM playback_history ph
         JOIN track t ON ph.track_id = t.id
