+++ web/src/lib/api/index.ts
@@ -73,9 +73,11 @@ export interface Artist {
 export interface Album {
     album: string;
     mbid?: string;
+    album_mbid?: string;
     art_id: number | null;
     art_sha1?: string | null;
     artist_name: string;
+    artist_mbid?: string;
     is_hires: number;
     year: string | null;
     track_count: number;
@@ -109,6 +111,9 @@ export interface Track {
     bit_depth: number | null;
     mb_release_id?: string | null;
     mb_release_group_id?: string | null;
+    artist_mbid?: string | null;
+    album_artist_mbid?: string | null;
+    album_mbid?: string | null;
     popularity?: number;
 }
 
@@ -136,16 +141,25 @@ export async function fetchArtists(
     return await res.json();
 }
 
-export async function fetchAlbums(params: { artist?: string; albumMbid?: string } = {}, fetchFn: any = fetch): Promise<Album[]> {
+export async function fetchAlbums(params: { artist?: string; artistMbid?: string; albumMbid?: string } = {}, fetchFn: any = fetch): Promise<Album[]> {
     let url = '/api/albums';
+    const query = new URLSearchParams();
+
     if (params.artist) {
-        url += `?artist=${encodeURIComponent(params.artist)}`;
+        query.append('artist', params.artist);
     }
-    // Support album MBID query
-    if ((params as any).albumMbid) {
-        const q = params.artist ? '&' : '?';
-        url += `${q}album_mbid=${encodeURIComponent((params as any).albumMbid)}`;
+    if (params.artistMbid) {
+        query.append('artist_mbid', params.artistMbid);
     }
+    if (params.albumMbid) {
+        query.append('album_mbid', params.albumMbid);
+    }
+
+    const queryString = query.toString();
+    if (queryString) {
+        url += `?${queryString}`;
+    }
+
     const res = await fetchFn(url);
     if (!res.ok) throw new Error('Failed to fetch albums');
     return await res.json();
